requirements.txt
pyexcel
pyexcel-xlsx
pyexcel-ods3
aiosmtplib
customtkinter
matplotlib
pillow